refactor: project sructure
