+++ src/handlers/hash-calculation.js
@@ -1,8 +1,8 @@
 import { readFile } from 'fs/promises';
 import { createHash } from 'crypto';
 
-import { errorCode } from './const.js';
-import { writeMessage } from './utils.js';
+import { errorCode } from '../const.js';
+import { writeMessage } from '../utils/input-output.js';
 
 export const hash = async (path) => {
   if (!path) throw new Error(errorCode.noUrl);
