+++ src/utils/cli.js
@@ -1,0 +1,5 @@
+export const getUsername = () => {
+  let result = process.argv.find((arg) => arg.includes('--username'));
+  if (result) result = result.slice(11);
+  return result;
+};
